cpu v1.1

+++ cpu/cpu.h
@@ -4,8 +4,8 @@
 #include <string.h>
 #include <unistd.h>
 
-double cpu_total_usage();
+float cpu_total_usage();
 unsigned long cpu_total_time();
 float cpu_usage_pid_in_5_min(const char *);
 void cpu_show_usage_pid_in_5_min(const char *, float);
-void cpu_show_usage(double);
+void cpu_show_usage(float);
